// frontend/src/app/models/Event.ts
class Event {
    name: string | undefined;
    chanel: string | undefined;
    startDate: string | undefined;
    endDate: string | undefined;
    partners: string | undefined;
    targets: string | undefined;
    participants: string | undefined;
    note: string | undefined;
    status: string | undefined;
  }
  export default Event;
  